+++ R/class-general.R
@@ -90,7 +90,7 @@ setAttributeColumn <- function(object, attr_name, col_or_value, col_name, defaul
   }
   # Apply col_or_value to col_name or to default_value
   if (!missing(col_or_value)) {
-    if (col_or_value %in% names(object[[dataname]])) {
+    if (all(col_or_value %in% names(object[[dataname]]))) {
       col_name <- col_or_value
     } else {
       default_value <- col_or_value
@@ -161,4 +161,41 @@ getAttributeColumn <- function(object, attr_name, warn_missing=c("attr", "column
   } else {
     object[[dataname]][, columns, drop=FALSE]
   }
+}
+
+#' When parsing the input for \code{PKNCAconc} and \code{PKNCAdose} 
+#' allow either formula or character inputs.
+#' 
+#' @param formula A formula with an optional left side, right side, and
+#'   groups after a pipe.
+#' @param lhs,rhs,groups Alternate definition of the formula
+#' @return A list with named elements for the "lhs", "rhs", and "groups"
+formulaOrNames <- function(formula=NA_character_,
+                           lhs=NA_character_, rhs=NA_character_, groups=c()) {
+  ret <- list()
+  if (!is.na(formula)) {
+    parsedFormula <- parseFormula(formula, require.two.sided=FALSE)
+    ret$lhs <- all.vars(parsedFormula$lhs)
+    ret$rhs <- all.vars(parsedFormula$rhs)
+    ret$groups <- all.vars(parsedForm$groupFormula)
+  } else {
+    ret$lhs <- lhs
+    ret$rhs <- rhs
+    ret$groups <- groups
+  }
+  if (identical(ret$lhs, ".")) {
+    ret$lhs <- NA_character_
+  } else if (length(ret$lhs) != 1) {
+    stop("The left side of the formula must have only one variable")
+  } else if (is.na(ret$lhs)) {
+    ret$lhs <- NA_character_
+  }
+  if (identical(ret$rhs, ".")) {
+    ret$rhs <- NA_character_
+  } else if (length(ret$rhs) != 1) {
+    stop("The right side of the formula must have only one variable")
+  } else if (is.na(ret$rhs)) {
+    ret$rhs <- NA_character_
+  }
+  ret
 }